Add files via upload
Project 3 Specifications
You have been tasked to write a Java program that will go through the first step of encrypting an input file and writing the encrypted contents to an output file.

For each letter read in from the input file, add 3 letters to the value of the input letter.  If a letter is uppercase in the input file, the encrypted letter should be converted to a lowercase character in the output file.  If a letter is lowercase in the input file, the encrypted letter should be converted to an uppercase character in the output file.  This means if the input letter is an A then the encrypted letter would be the letter d.  If the input letter from the input file is an X then the encrypted letter would be the letter a.

All spaces, digits, punctuation, and special characters do not need to be encrypted at this level.   Simply write the space, punctuation, or special character to the output file as is.

Since the input file will be of various lengths, the program will read until end of file.

The input file should be located in the current directory and is named Oct1.txt.   The output file will be written to the current directory and named EncryptOct1.txt. 

#Encryption